build: update to latest @rollup/plugin-commonjs version
Updates the project to the latest version of `@rollup/plugin-commonjs`. For
this, `rollup` also needs to be updated as the peer dependency from the
CommonJS plugin would not match otherwise.

This results in a change in golden files as the rollup output is slightly
different. We update the goldens as part of this commit.

The update to the latest CommonJS rollup plugin version is in preparation
of bundling the runfile helpers into a checked-in file. Since the runfile
helpers use a dynamic import for loading the NodeJS patches script, the
CommonJS rollup plugin needs to respect those. Currently the CommonJS plugin
just invalidates the dynamic require, while the newer version allows us to
preserve the dynamic CommonJS import as is.