chore: cleanup version checking code
We can just use the bazel-skylib helper to check the Bazel version.
Also we haven't used the check_rules_nodejs_version helper in forever, so delete the implementation and make a note to remove it in 4.0.
This also fixes a bug where our version parser didn't handle semver 'build metadata' which is like +something on the end