fix(builtin): include optionalDependencies in strictly visible packages
These are identical to regular dependencies except that a failure to install them is non-fatal.
They should be visible as Bazel dependencies.

Note I didn't include peerDependencies here still. These are an indication that some consumer should install the package, and that results in a bazel-visible dependency, so I think that's working correctly