fix(builtin): generated nodejs repository for windows references non-existent file
The generated NodeJS repository for windows. e.g. `@nodejs_windows_amd64`
references a non-existent file that is supposed to be the Yarn script.

The `yarn.cmd` file is located under `bin/` in the extracted Yarn
archive, and not at the top-level. This currently breaks integration
tests in the Angular framework due to the following error when
the following target is referenced: `@nodejs//:yarn_bin`.
```
 missing input file
 'external/nodejs_windows_amd64/bin/yarnpkg/yarn.cmd', owner:
 '@nodejs_windows_amd64//:bin/yarnpkg/yarn.cmd'
```