feat: allow esbuild cwd for subprocess to be set
This change is to allow users to set where the esbuild process is
invoked relative to the root of the repo. This is useful in the case
of monorepos in which packages may be nested and not invoked from
root. This is important to have as it helps with relative path resolution
when an entry_point is passed ot ESBuild. Without this, ESBuild will fail
to resolve relative imports.