refactor: convert web_testing example to ts_project
We no longer recommend using ts_library so our examples shouldn't demonstrate it.
Note that for karma_web_test to work, it needs the JSNamedModuleInfo provider, so each ts_project is wrapped with a js_library that exposes the .js files to that provider. Also we add the triple-slash directive so TypeScript produces named umd outputs that work with concatjs