feat(cypress): cypress executable toolchain (#2668)
This is a major refactor the cypress_repository rule
Prior to this change, the cypress CLI from NPM was invoked
to download the corresponding Cypress binary.

This was bad for a number of reasons. Namely it was non cachable
and did not match bazel's semantics for grabbing the right binary
for a particular platform / architecture

The new cypress_repositories rule, downloads binaries for all
platforms from Cypress's CDN and uses bazel's select logic
to use the appropriate one given a particular configuration.