chore: remove version checking from @bazel/worker
This package is meant to be a utility required by code outside our repo, on its own release cadence.
For example, rules_sass has a sass worker program that depends on @bazel/worker.

This means that new major releases of rules_nodejs will likely be used with an older version of @bazel/worker.
We intend for the worker package to be quite stable across releases, so the check for matching major versions just makes upgrades very difficult, with no benefit.

Fixes #2807