fix: update usages of `ExternalNpmPackageInfo.path` to be safe and default to empty string
`ExternalNpmPackageInfo` had a `path` attribute introduced and used in
2c2cc6eec42b0f4f805fb3063920107b78f821fd, but the provider was publicly exported beforehand. Even
though all internal usages were updated to include a path, this change was shipped in `3.2.0` and
introduced an unintentional breaking change for consumers. Most notably, `rules_postcss` manually
implements this provider and does not give a `path` failing in versions `3.2.0` and up.

https://github.com/bazelbuild/rules_postcss/blob/2bd16fda40cd4bf4fbf0b477b968366ec1602103/internal/plugin.bzl#L30-L41

Solution here is to safely retrieve the `path` attribute and default to empty string at all places
it is used. A test is also introduced to ensure that a `nodejs_binary()` can depend on a custom
`ExternalNpmPackageInfo` provider without a `path` attribute, which should hopefully prevent against
regressions.