Fixed ts_project json file outpath with root_dir
When using the root_dir attr, the paths to json files are not modified from the input file paths in the same fashion that ts files are. I updated the definition of json_outs to use the same logic to produce an out file path that is used with ts files.