refactor: move node_patches.js to core
This allows us to wire it into new implementation for nodejs_binary

Note that the sources for this are still awkwardly in the packages/
folder despite us not publishing such a package to npm.

This is okay since the core package doesn't depend on these sources,
they're updated through a diff_test/update pattern.
Moving the sources is harder because the core doesn't even have a way to
compile TypeScript yet, and ideally it never will.