test: e2e/core shows nodejs_binary, npm_import
Currently the core package only exposes a nodejs toolchain. It contains just the code in the /nodejs folder which is our new cleanroom (it doesn't depend outside that folder)

Introduces three rules, currently in the e2e/core package while we decide about the shape before promoting them to the core package

    npm_import fetches an npm package using Bazel's downloader (so it's cached outside the repository cache). It exposes a TreeArtifact as the default output, though this will need to be some kind of *_library before it can really be used
    nodejs_binary is a bare-bones clean implementation, no bash required on windows, and when the runfiles are enabled it lays out a node_modules folder with acorn package. Right now that package is hard-coded into the spots
    nodejs_test is just the same as nodejs_binary with test = True rather than executable = True

This is a blueprint for what we might want to push up into the core package so that it can host trivial nodejs programs with third-party dependencies.

Later work:

    nodejs_library as the thing npm_import exposes
    there's a TODO to generate the WORKSPACE section from a package-lock.json file so users don't need to write npm_import rules
    a linker so that programs can resolve third-party dependencies that weren't in the runfiles of the nodejs_binary they call
    more starlark testing of things in the /nodejs package