feat(esbuild): make Starlark build settings usable as defines (#3122)
Adds a `define_settings` attribute to the `esbuild` rules that allows
using the values of Starlark build settings to globally replace
specified identifiers.

The value of the setting is automatically converted to a JS literal.
Note: This currently uses json.encode, which is only available from
Bazel 5 on. If this addition is generally viewed favorably, I would
look into "polyfilling" this functionality in Starlark.