Revert "fix(builtin): add transitive typings to runfiles provider produced by js_library (#2547)"
This wasn't actually a fix - it causes type-checking to happen eagerly
when a js_library target is built, even if the typings aren't needed.

This prevents a transpile-only dev workflow.

As noted in the PR https://github.com/bazelbuild/rules_nodejs/pull/2547
this is easily accomplished in tsc_test by explicitly gathering typings
inputs with output_group = types

This reverts commit 41117fa93fecde3563482a9fac54073a7b529911.