fix(cypress): patch cypress to not run verify
When cypress first runs, it creates a binary_state.json file to
indicate that it has tested out the downloaded test runner before.
However, since bazel runs in a sandbox, every test exectution is the
first test run. We don't want to waste time verifying so we instead
mock the behavior of cypress by writing the file ourselves.