refactor: promote directory_file_path to public API
Users with custom rules need to access this when the rule wants to reference labels within a tree artifact
such as when using exports_directories_only.

Note, I didn't leave anything behind in the private/ folder since the rule implementation is just one line.

This also adds the rule to the documentation, previously only the DirectoryFilePathInfo was documented.