fix(jasmine): replace deprecated Jasmine APIs that have been removed in version 4
With this change we replace APIs that have been deprecated in version 4 and removed in version 4

```
Jasmine#onComplete is deprecated. Instead of calling onComplete, set the Jasmine instance's exitOnCompletion property to false and use the promise returned from the execute method.
```

Also this addresses the breaking change in version 4 https://github.com/jasmine/jasmine/blob/main/release_notes/4.0.0.md#changes-that-affect-custom-reporters in a backward compatible manner.

Closes #3289