fix(builtin): `yarn_install` with vendored yarn `.cjs` file breaks
`yarn_install` allows for a vendored/custom Yarn JavaScript file. To ensure
that Yarn runs with the proper Node version, `yarn_install` calls the JS file
with the host NodeJS version also installed through Bazel. Currently the detection
logic does not account for `.cjs` or `.mjs`. This is a little confusing as the
install continues to work but breaks surprisingly in Windows then, or when a engines
requirement is no longer satisifed (there are more reasons as well)

We should support `cjs` and `mjs` to avoid confusion, and especially `cjs`
since Yarn itself always vendors itself with that extension!