test: update environment spec to work with windows
It looks like the environment spec started failing on Windows. It was
thought that this started happening with 4a025c0d197a181cfcd9556831f69467817d3a40,
but it turnes out that this failed many commits before but a Buildkite status
just did not show up in the commit history (making it look like the PnP commit
caused the failures).