refactor(builtin): allow static files to be specified in `pkg_npm` `create_package` helper
The `pkg_npm` builtin exposes a `create_package` helper that is leveraged
by the `ng_package` Angular Bazel rule. The `create_package` helper
currently always determines files from the `ctx`, and it's not possible
to add more files to the package, or to omit files that are already
processed by the `ng_package` wrapping rule implementation.

The static files should be configurable similar to deps and the nested
packages.