Teach os_name() about freebsd-amd64
It is possible to use `rules_nodejs` on operating systems besides
Windows, Linux, and macOS, but they need to be handled in `os_name()`
or else you get a `Unsupported operating system freebsd` error message.
This change allows you to add a FreeBSD toolchain to your project by
declaring a `nodejs_freebsd_amd64` repo with a `bin/node` executable.
cockroachdb/cockroach#82762 is an example of this.