feat: provide @nodejs repository
Generally it's a bad practice that the user calls a repository rule (nodejs_register_toolchains) with a name, and this doesn't result in a repository with that name.

Leave the old _host variant around to make this a non-breaking change.

We did the same fix in python: https://github.com/bazelbuild/rules_python/pull/656

Fixes #3375