chore: bzlmod
Previous rules_nodejs entries on bazel-central-registry were created by hand.
This moves our 'toolchain' example which is the natural e2e for bare rules_nodejs.
We test it on BCR in the normal way.
I'll need to make sure the Publish To BCR app is enabled for the bazelbuild/rules_nodejs repo.