Work properly on intel CPUs. Normalize x86_64->amd64 naming.
In particular on x86 macbooks, the rctx.os.arch returns x86_64 (and not 'amd64')
as expected by the nodejs namings.

This led to errors like:

```
... depends on @nodejs-macos-x86_64//:node_toolchain in repository @nodejs-macos-x86_64 which failed to fetch. no such package '@nodejs-macos-x86_64//': Unsupported operating system darwin architecture x86_64
```
The reason was the node_repositories role tries to check whether there is node for the host system on the list (BTW: I think this check is over-sensitive),
on was not recognizing the x86_64 architecture as supported.